Refactor entire folder structure > add README to each folder
