+++ src/app/app.routes.js
@@ -1,17 +1,8 @@
 const app = require("./app.middleware");
+const routes = require("../routes");
 
-app.get("/", async (req, res) => {
-  res.send({
-    message: "Storefront API. See documentation for use.",
-  });
-});
+app.use('/', routes.app)
+app.use("/api/tasks", routes.tasks);
 
-app.use("/products", require("../routes/productsRoutes"));
-app.use("/logos", require("../routes/logosRoutes"));
-app.use("/stores", require("../routes/storesRoutes"));
-app.use("/categories", require("../routes/categoriesRoutes"));
-app.use("/product-types", require("../routes/product-typesRoutes"));
-app.use("/variations", require("../routes/variationsRoutes"));
-app.use("/api/tasks", require("../routes/tasksRoutes"));
 
 module.exports = app;
