+++ src/app/index.js
@@ -1,3 +1,6 @@
 const app = require("./app.routes");
 
+// we export it from an index file to avoid redundancy in /app/app.js
+// now we can do /app/ when importing it
+
 module.exports = app;
