+++ src/database/index.js
@@ -1,0 +1,7 @@
+const database = {
+  info: require("./system.info"),
+  functions: require("./system.functions"),
+  config: require("./system.config"),
+};
+
+module.exports = database;
