--- src/http_tests/categories.http
@@ -1,27 +1,0 @@
-### Test the categories
-### categories are the labels under which products are organized
-
-GET http://localhost:3001/categories
-
-### Test creating a category after you build the code for adding a category
-POST http://localhost:3001/categories
-Content-Type: application/json
-
-{
-  "category": "Men's Apparel"
-}
-
-### Test the PUT which should change a category
-PUT http://localhost:3001/categories/5f30519cdd1f2c20488d027c
-Content-Type: application/json
-
-{
-  "category": "Unisex Apparel"
-}
-
-### Test the DELETE which should remove a logo
-
-
-DELETE http://localhost:3001/categories/5f30519cdd1f2c20488d027c
-
-
