--- src/http_tests/logos.http
@@ -1,26 +1,0 @@
-### Test the home page endpoint
-
-GET http://localhost:3001/logos
-
-### Test creating a logo after you build the code for adding a logo
-POST http://localhost:3001/logos
-Content-Type: application/json
-
-{
-  "name": "Titleist",
-  "descirption": "As a child i thought this was pronounced Tit Leist",
-  "colors": ["Black", "Red"]
-}
-
-### Test the PUT which should change a logo
-PUT http://localhost:3001/logos/5f2c919f77f3d857d0566ee6
-Content-Type: application/json
-
-{
-  "name": "Super Logo"
-}
-
-### Test the DELETE which should remove a logo
-
-
-DELETE http://localhost:3001/logos/5f2c920ed615b533c42c9d0d
