--- src/http_tests/product-types.http
@@ -1,29 +1,0 @@
-### Test the types
-### types are the product types
-# Product types (or product classes) are groups of products which share the same attributes.
-
-# Product attributes contain additional product information, e.g. ISBN, UPC, Brand, which is displayed in storefront and included in product feeds when exporting to marketplaces like Google Shopping, eBay, Amazon ads etc.
-
-GET http://localhost:3001/product-types
-
-### Test creating a type
-
-POST http://localhost:3001/product-types
-Content-Type: application/json
-
-{
-    "Type": "Polo Shirt"
-}
-
-### Test the PUT which should change a product type
-
-PUT  http://localhost:3001/product-types/5f30598df68fc848702a50f5
-Content-Type: application/json
-
-{
-    "Type": "Yolo Shirt"
-}
-
-### Test the DELETE which should delete a product type
-
-DELETE http://localhost:3001/product-types/5f30598df68fc848702a50f5
