--- src/http_tests/products.http
@@ -1,25 +1,0 @@
-### Test the home page endpoint
-
-GET http://localhost:3001/products
-
-### Test creating a product
-POST http://localhost:3001/products
-Content-Type: application/json
-
-{
-  "title": "Beer Coozie",
-  "description": "The most important part of any round of golf"
-}
-
-### Test the PUT which should change a product
-### id of 5f2b54de320478eb4b1603eb is temporary while server is running
-PUT http://localhost:3001/products/5f2c7a7832d9964f9d97826e
-Content-Type: application/json
-
-{
-  "title": "Amazing UV Shirt"
-}
-
-### Test the DELETE which should delete a product
-DELETE  http://localhost:3001/products/5f2c7a7832d9964f9d97826e
-
