--- src/http_tests/variations.http
@@ -1,25 +1,0 @@
-### Test the variations
-### variations are the different styles that can be applied to a product
-
-GET http://localhost:3001/variations
-
-### Test creating a variation
-
-POST http://localhost:3001/variations
-Content-Type: application/json
-
-{
-    "variation": "Red"
-}
-
-### Test the PUT which should change a variation
-
-PUT http://localhost:3001/variations/5f305ed5a159c55254bb16d5
-Content-Type: application/json
-
-{
-    "variation": "Dark Red"
-}
-
-### Test DELETE which should delete a variation
-DELETE http://localhost:3001/variations/5f305ed5a159c55254bb16d5
