+++ src/routes/index.js
@@ -1,0 +1,6 @@
+const routes = {
+    tasks: require('./tasks.routes'),
+    app: require('./app.routes')
+}
+
+module.exports = routes
